+++ example/lib/src/view.dart
@@ -20,11 +20,9 @@
 ///          Created  16 Dec 2018
 ///
 
-
 export 'package:mvc_application/view.dart';
 
 export 'view/app/contacts_example_app.dart';
 export 'view/home/add_contact.dart';
-export 'view/home/address_tile.dart';
 export 'view/home/contact_details.dart';
 export 'view/home/contacts_list.dart';